Sonic Trilogy: Remastered Launcher
This was intended to be a launcher for Sonic 1,2,and 3AIR that would be more than just a simple launcher.

It would give info on the games, their prototypes, cheats, unused content and even interviews!

Sadly due to being built in game maker it  is unable to preform it's core function- launching the games.

That being said someone may beable to translate this to C or C++

If you do keep the overall interface intac.t

I'll upload the resources to deviantart, as there are parts of the interface that haven't been implemented yet.

Those can only be found there.