C++ Update (REVIVAL, Unstable build)
The Launcher has moved to C++ as GameMaker Studio 2 was not up to the challenge of making the Launcher a reality.

Luckily C++ was very willing to help in making it a reality.

However I know very little about C++ so I'll have to learn, for now the launcher is in a VERY basic state so keep that in mind.

Basically You won't get this:
<IMAGE1>
You'll get this:
<IMAGE2>

On the bright side here's what I have done:
+ Games boot as they should (Sonic 1 and 2, are now getting the data they need)
+Sonic 3 AIR now boots.
+As said before you'll be asked to put the location of the game in.

Current Issues (For me specifically):
- Doesn't remember the User's input path.

Fix: The path is saved to a variable, so save the variable to a file then get that varible from the file.

-Once a game is closed so does the launcher.

-No Ideas yet on how to add the interface (Somone suggested SDL, so i figure that.)

To Be Implemented:
-Extras Menu not  Implemented
-No Graphics (Yet).
-Extras Menu's Submenus are not implemented.
-No info.