GameMaker Build
This is the old build made in GameMaker Studio 2.

It unlike the C version has the following:
-Games Boot
-Interface Implemented.
-Menu controlled via mouse.

Issues:
-Working Directoriy changing

Explanation:
Sonic 1 and 2 (2013 decomps) use the same program: RSDKv4/v4_64.exe

Because of this they only differ in the files being loaded.

Due to this without the ability to change the working directory, the games WILL boot but fail to load the files they need.

Trying to provide them in the default working directory will cause them to boot but Sonic 1 and 2's files cannot coexist.

This is due to the Level files:
Level01,Level02,03...etc,

Windows will ask you if you want to replace them- obviously not.

Unless you want to play broken versions of the games.

Sonic 3 AIR I theorize has the same issue.

If anyone wants to try and get this version working - go ahead.

I personally want to stick with C++ as it will help me in the future to know it.

TL;DR The games fight over files.

-Mouse Control.
Menu is controlled by mouse for controller support a data structure grid (ds_grid) is required.

-GameMaker Studio 2 is not up to the task.

Don't know why I thought it was, it's called Game Maker for a reason, it's built to make Video games, not launch them.

Hence why I switched to C++